filtroneSite/requirements.txt
absl-py==0.15.0
apturl==0.5.2
asgiref==3.4.1
astunparse==1.6.3
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.9
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
flatbuffers==2.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.5.0
httplib2==0.14.0
idna==2.8
joblib==1.1.0
kazam==1.4.5
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==12.0.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
numpy==1.21.4
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.3.4
paramiko==2.6.0
pexpect==4.6.0
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.2
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
systemd-python==234
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
threadpoolctl==3.0.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==2.0.2
wrapt==1.13.3
xkit==0.0.0
